requirements.txt
asttokens==2.0.5
backcall==0.2.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
ipykernel==6.13.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.1.0
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.2.2.post1
wcwidth==0.2.5
